Refactored Prechelts_test_program.py
Upon reviewing discussion of the Problem by Peter Norvig and then
carefully re-reading the problem statement, I found that I had made
an incorrect assumption about exactly when a digit could be inserted
into the encoding. A slight change in logic was required to get the
right logic. Tests pass.